backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0
lightgbm==4.3.0
joblib==1.3.2

# OpenAI Integration
openai==1.10.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
